Flush cache when lock/unlock cache ways in scheduler
PL310 only control the cache allocation. Once content has been loaded
into the cache, it is out of PL310's control.

If we do not flush cache when lock/unlock cache ways in scheduler, we
will not get the cache isolation as we want. The cache pollution task
may load its content into the whole cache, interfere other tasks and
cause the RT task under GSN-FPCA has a larger WCET under interference
than under GSN-FP.

GSN-FPCA outperforms the GSN-FP under interference when we flush the
cache ways explicitly at locking/unlocking the cache ways.

Test
    WCET of RT task under interference is same with the WCET of RT when
it runs alone under GSN-FPCA algorithm

+++ arch/arm/mm/cache-l2x0.c
@@ -148,7 +148,7 @@ static void __l2x0_flush_all(void)
 	debug_writel(0x00);
 }
 
-void l2x0_flush_cache_ways(u32 ways)
+void l2x0_flush_cache_ways(uint16_t ways)
 {
 	unsigned long flags;
 	ways &= l2x0_way_mask;
