+++ include/litmus/cache_proc.h
@@ -12,7 +12,7 @@ int unlock_cache_ways_to_cpu(int cpu);
 int __unlock_cache_ways_to_cpu(int cpu);
 int __get_used_cache_ways_on_cpu(int cpu, uint16_t *cp_mask);
 
-void l2x0_flush_cache_ways(u32 ways);
+void l2x0_flush_cache_ways(uint16_t ways);
 
 #endif
